+++ R/data.R
@@ -345,6 +345,7 @@
 #' \describe{
 #' \item{Re_dempl}{Direct emplacement effect rebound, always 0.}
 #' \item{Re_emb}{Indirect embodied energy effect rebound.}
+#' \item{Re_cap}{Indirect capital cost effect rebound.}
 #' \item{Re_md}{Indirect maintenance and disposal effect rebound.}
 #' \item{Re_empl}{Emplacement effect rebound.}
 #' \item{Re_dsub}{Direct substitution effect rebound.}
@@ -363,6 +364,23 @@
 "rebound_terms"
 
 
+#' Aggregate rebound terms
+#' 
+#' This is the list of the rebound terms that are aggregates of other rebound terms.
+#' 
+#' @format A string list with `r length(rebound_terms_agg)` entries.
+#' \describe{
+#' \item{Re_empl}{Emplacement effect rebound.}
+#' \item{Re_sub}{Substitution effect rebound.}
+#' \item{Re_inc}{Income effect rebound.}
+#' \item{Re_d}{Sum of all direct rebound effects.}
+#' \item{Re_i}{Sum of all indirect rebound effects.}
+#' \item{Re_tot}{Total rebound.}
+#' }
+#' @examples
+#' rebound_terms_agg
+"rebound_terms_agg"
+
 #' LaTeX version of rebound terms
 #'
 #' This is the list of the rebound terms in LaTeX form.
@@ -371,6 +389,7 @@
 #' \describe{
 #' \item{Re_dempl}{Direct emplacement effect rebound, always 0.}
 #' \item{Re_emb}{Indirect embodied energy effect rebound.}
+#' \item{Re_cap}{Indirect capital cost effect rebound.}
 #' \item{Re_md}{Indirect maintenance and disposal effect rebound.}
 #' \item{Re_empl}{Emplacement effect rebound.}
 #' \item{Re_dsub}{Direct substitution effect rebound.}
@@ -428,6 +447,10 @@
 #' \item{sigma_col}{The name of the column containing ending sigma values for this indifference curve.}
 #' \item{start_point_col}{The name of the boolean column telling whether this row contains a segment that should have a starting point.}
 #' \item{end_arrow_col}{The name of the boolean column telling whether this row contains a segment that should have an ending arrow.}
+#' \item{Re_names}{The name of the column of rebound names.}
+#' \item{Re_values}{The name of the column of rebound values.}
+#' \item{y_names_col}{The name of the column of y-axis variable names.}
+#' \item{y_vals_col}{The name of the column of y-axis variable names.}
 #' }
 #' @examples
 #' graph_df_colnames
@@ -459,31 +482,46 @@
 #' \item{emb_colour}{The colour for embodied energy lines.}
 #' \item{cap_colour}{The colour for capital cost lines.}
 #' \item{md_colour}{The colour for maintenance and disposal lines.}
+#' \item{empl_colour}{The colour for emplacement lines.}
 #' \item{dsub_colour}{The colour for direct substitution lines.}
 #' \item{isub_colour}{The colour for indirect substitution lines.}
+#' \item{sub_colour}{The colour for substitution lines.}
 #' \item{dinc_colour}{The colour for direct income lines.}
 #' \item{iinc_colour}{The colour for indirect income lines.}
+#' \item{inc_colour}{The colour for income lines.}
 #' \item{prod_colour}{The colour for productivity lines.}
+#' \item{dir_colour}{The colour for direct lines.}
+#' \item{indir_colour}{The colour for productivity lines.}
 #' \item{tot_colour}{The colour for total rebound lines.}
 #' \item{dempl_size}{The size for direct emplacment lines.}
 #' \item{emb_size}{The size for embodied energy lines.}
 #' \item{cap_size}{The size for capital cost lines.}
 #' \item{md_size}{The size for maintenance and disposal lines.}
+#' \item{empl_size}{The size for emplacement lines.}
 #' \item{dsub_size}{The size for direct substitution lines.}
 #' \item{isub_size}{The size for indirect substitution lines.}
+#' \item{sub_size}{The size for substitution lines.}
 #' \item{dinc_size}{The size for direct income lines.}
 #' \item{iinc_size}{The size for indirect income lines.}
+#' \item{inc_size}{The size for income lines.}
 #' \item{prod_size}{The size for productivity lines.}
+#' \item{dir_size}{The size for direct lines.}
+#' \item{indir_size}{The size for indirect lines.}
 #' \item{tot_size}{The size for total rebound lines.}
 #' \item{dempl_linetyps}{The linetype for direct emplacment lines.}
 #' \item{emb_linetype}{The linetype for embodied energy lines.}
 #' \item{cap_linetype}{The linetype for capital cost lines.}
 #' \item{md_linetype}{The linetype for maintenance and disposal lines.}
+#' \item{empl_linetype}{The linetype for emplacement lines.}
 #' \item{dsub_linetype}{The linetype for direct substitution lines.}
 #' \item{isub_linetype}{The linetype for indirect substitution lines.}
+#' \item{sub_linetype}{The linetype for substitution lines.}
 #' \item{dinc_linetype}{The linetype for direct income lines.}
 #' \item{iinc_linetype}{The linetype for indirect income lines.}
+#' \item{inc_linetype}{The linetype for income lines.}
 #' \item{prod_linetype}{The linetype for productivity lines.}
+#' \item{dir_linetype}{The linetype for direct lines.}
+#' \item{indir_linetype}{The linetype for indirect lines.}
 #' \item{tot_linetype}{The linetype for total rebound lines.}
 #' \item{lineend}{The line end style.}
 #' \item{linejoin}{The line join style.}
@@ -546,44 +584,62 @@
 #' @format A string list with `r length(sens_graph_params)` entries.
 #' \describe{
 #' 
-#' \item{orig_point_shape}{The shape for points between rebound effects. Default is `21`, a filled circle..} 
-#' \item{orig_point_size}{The size for points between rebound effects. Default is `1`.}
-#' \item{orig_point_stroke}{The size of the line surrounding points between rebound effects. Default is `1`.}
-#' \item{orig_point_colour}{The colour of the original points. Default is "red".}
+#' \item{orig_point_shape}{The shape for points between rebound effects. Default is `1`, an open circle..} 
+#' \item{orig_point_size}{The size for original points. Default is `2`.}
+#' \item{orig_point_stroke}{The size of the line surrounding points between rebound effects. Default is `0.5`.}
+#' \item{orig_point_colour}{The outline colour for original points. Default is "black".}
+#' \item{orig_point_fill}{The fill colour for original points. Default is "black".}
 #' \item{dempl_colour}{The colour for direct emplacment lines.}
 #' \item{emb_colour}{The colour for embodied energy lines.}
 #' \item{cap_colour}{The colour for capital cost lines.}
 #' \item{md_colour}{The colour for maintenance and disposal lines.}
+#' \item{empl_colour}{The colour for emplacement lines.}
 #' \item{dsub_colour}{The colour for direct substitution lines.}
 #' \item{isub_colour}{The colour for indirect substitution lines.}
+#' \item{sub_colour}{The colour for substitution lines.}
 #' \item{dinc_colour}{The colour for direct income lines.}
 #' \item{iinc_colour}{The colour for indirect income lines.}
+#' \item{inc_colour}{The colour for income lines.}
 #' \item{prod_colour}{The colour for productivity lines.}
+#' \item{dir_colour}{The colour for direct lines.}
+#' \item{indir_colour}{The colour for productivity lines.}
 #' \item{tot_colour}{The colour for total rebound lines.}
 #' \item{dempl_size}{The size for direct emplacment lines.}
 #' \item{emb_size}{The size for embodied energy lines.}
 #' \item{cap_size}{The size for capital cost lines.}
 #' \item{md_size}{The size for maintenance and disposal lines.}
+#' \item{empl_size}{The size for emplacement lines.}
 #' \item{dsub_size}{The size for direct substitution lines.}
 #' \item{isub_size}{The size for indirect substitution lines.}
+#' \item{sub_size}{The size for substitution lines.}
 #' \item{dinc_size}{The size for direct income lines.}
 #' \item{iinc_size}{The size for indirect income lines.}
+#' \item{inc_size}{The size for income lines.}
 #' \item{prod_size}{The size for productivity lines.}
+#' \item{dir_size}{The size for direct lines.}
+#' \item{indir_size}{The size for indirect lines.}
 #' \item{tot_size}{The size for total rebound lines.}
 #' \item{dempl_linetyps}{The linetype for direct emplacment lines.}
 #' \item{emb_linetype}{The linetype for embodied energy lines.}
 #' \item{cap_linetype}{The linetype for capital cost lines.}
 #' \item{md_linetype}{The linetype for maintenance and disposal lines.}
+#' \item{empl_linetype}{The linetype for emplacement lines.}
 #' \item{dsub_linetype}{The linetype for direct substitution lines.}
 #' \item{isub_linetype}{The linetype for indirect substitution lines.}
+#' \item{sub_linetype}{The linetype for substitution lines.}
 #' \item{dinc_linetype}{The linetype for direct income lines.}
 #' \item{iinc_linetype}{The linetype for indirect income lines.}
+#' \item{inc_linetype}{The linetype for income lines.}
 #' \item{prod_linetype}{The linetype for productivity lines.}
+#' \item{dir_linetype}{The linetype for direct lines.}
+#' \item{indir_linetype}{The linetype for indirect lines.}
 #' \item{tot_linetype}{The linetype for total rebound lines.}
 #' \item{lineend}{The line end style.}
 #' \item{linejoin}{The line join style.}
 #' \item{include_base_condition_points}{Tells whether to include base conditions points. Default is `TRUE`.}
 #' \item{points_atop_paths}{Tells whether to draw points above paths (`TRUE`) or beneath paths (`FALSE`). Default is `TRUE`.}
+#' \item{use_latex_legend}{Tells whether to convert legend text to LaTeX for a nicely-formatted legend. Default is `FALSE`.}
+#' \item{include_x_axis}{Tells whether to add an x-axis at y = 0. Default is `FALSE`.}
 #' }
 #' @examples
 #' sens_graph_params