+++ R/graphs.R
@@ -22,15 +22,15 @@
 #'   rebound_analysis() %>% 
 #'   path_graphs(indexed = TRUE)
 path_graphs <- function(.analysis_data,
-                           indexed = FALSE,
-                           cases = .analysis_data[[case_colname]] %>% unique(),
-                           graph_types = ReboundTools::graph_types,
-                           grid_types = ReboundTools::graph_types,
-                           graph_params = ReboundTools::path_graph_params, 
-                           case_colname = ReboundTools::eeu_base_params$case, 
-                           rebound_stages = ReboundTools::rebound_stages,
-                           rebound_segments = ReboundTools::rebound_segments,
-                           graph_df_colnames = ReboundTools::graph_df_colnames) {
+                        indexed = FALSE,
+                        cases = .analysis_data[[case_colname]] %>% unique(),
+                        graph_types = ReboundTools::graph_types,
+                        grid_types = ReboundTools::graph_types,
+                        graph_params = ReboundTools::path_graph_params, 
+                        case_colname = ReboundTools::eeu_base_params$case, 
+                        rebound_stages = ReboundTools::rebound_stages,
+                        rebound_segments = ReboundTools::rebound_segments,
+                        graph_df_colnames = ReboundTools::graph_df_colnames) {
   
   cases <- match.arg(cases, several.ok = TRUE)
   graph_types <- match.arg(unlist(graph_types), choices = unlist(graph_types), several.ok = TRUE)
@@ -306,22 +306,35 @@ rebound_graphs_helper <- function(.path_data,
 
 #' Sensitivity graphs for rebound analyses
 #' 
-#' This function makes sensitivity graphs for rebound analysis.
+#' A function to make sensitivity graphs for rebound analysis.
 #' 
 #' The caller can adjust the aesthetics of the graph with manual scales.
+#' 
+#' The order of `y_var` will determine layering on the graph.
+#' The first item in `y_var` will be the lowest layer.
+#' The last item in `y_var` will be the highest layer.
 #'
 #' @param .parametric_data A data frame, likely the result of calling `parametric_analysis()`.
 #'                         Default is `parametric_analysis(rebound_data, parameterization)`.
 #' @param rebound_data Rebound data, likely read by `load_eeu_data()`.
+#'                     Default is `NULL`.
+#' @param parameterization A list of lists that gives parameter sweeps.
+#'                         At the top level, the list items must be named for cases in `rebound_data`.
+#'                         At the next level, the parameters to be swept should be given, 
+#'                         along with their sweep values. 
+#'                         See examples.
+#'                         Default is `NULL`.
 #' @param x_var,y_var Strings that identify the x-axis and y-axis variables for this sensitivity graph.
 #'                    These variables must appear in `.parametric_data`.
-#' @param linetype_var,linecolour_var,linesize_var Strings that identify variables to be used for
-#'                                                 type, color, and size of lines.
-#'                                                 Default is `ReboundTools::eeu_base_params$case`.
+#'                    `x_var` must be a single string.
+#'                    `y_var` can be a vector of strings. 
+#'                    See examples.
+#' @param line_var The name of variable to be used to discriminate lines on the graph. 
+#'                 Default is `y_names_col`.
 #' @param graph_params A list of parameters to control graph appearance. 
 #'                     See `ReboundTools::sens_graph_params`.
+#' @param y_names_col,y_vals_col See `ReboundTools::graph_df_colnames`.
 #' @param point_type_colname,sweep_points,orig_points See `ReboundTools::parametric_analysis_point_types`.
-#' @inheritParams parametric_analysis
 #' 
 #' @return A `ggplot2` object.
 #' 
@@ -334,9 +347,15 @@ rebound_graphs_helper <- function(.path_data,
 #'                     Lamp = list(k = seq(0, 2, by = 1)))
 #' sensitivity_graphs(rebound_data = df, parameterization = sens_params, 
 #'                    x_var = "k", y_var = "Re_tot") +
-#'  ggplot2::scale_colour_manual(values = c(Car = "black", Lamp = "black")) + 
-#'  ggplot2::scale_size_manual(values = c(Car = 0.5, Lamp = 0.5)) + 
-#'  ggplot2::scale_linetype_manual(values = c(Car = "solid", Lamp = "dashed"))
+#'  ggplot2::facet_wrap(facets = "Case", scales = "free_x") +
+#'  ggplot2::scale_colour_manual(values = c(Re_tot = "black"), guide = FALSE) + 
+#'  ggplot2::scale_size_manual(values = c(Re_tot = 0.5), guide = FALSE) + 
+#'  ggplot2::scale_linetype_manual(values = c(Re_tot = "solid"), guide = FALSE) +
+#'  ggplot2::labs(y = expression(Re[tot]), 
+#'                colour = ggplot2::element_blank(),
+#'                size = ggplot2::element_blank(),
+#'                linetype = ggplot2::element_blank())
+#'                
 #' # A more-complicated example that shows multi-variate sensitivity.
 #' # Values of the productivity parameter (k) is shown in rows of the lattice plot.
 #' # Uncompensated price elasticity of energy service consumption (e_qs_ps_UC) 
@@ -350,56 +369,251 @@ rebound_graphs_helper <- function(.path_data,
 #'                     Lamp = list(k = seq(0, 2, by = 0.5),
 #'                                 I_E = seq(2, 5, by = 1), 
 #'                                 e_qs_ps_UC = seq(-0.5, -0.1, by = 0.1)))
+#' # Choose which rebound variables to include and their order.
 #' sensitivity_graphs(rebound_data = df, parameterization = sens_params_2, 
-#'                    x_var = "I_E", y_var = "Re_tot") +
+#'                    x_var = "I_E", 
+#'                    y_var = "Re_tot",
+#'                    line_var = "Case") +
 #'   ggplot2::facet_grid(rows = ggplot2::vars(k), 
 #'                       cols = ggplot2::vars(e_qs_ps_UC), scales = "free_y") +
-#'   ggplot2::scale_colour_manual(values = c(Car = "darkgreen", Lamp = "black")) + 
-#'   ggplot2::scale_size_manual(values = c(Car = 0.5, Lamp = 1)) + 
-#'   ggplot2::scale_linetype_manual(values = c(Car = "solid", Lamp = "dotted")) + 
+#'   ggplot2::scale_colour_manual(values = c(Car = "black", Lamp = "red")) + 
+#'   ggplot2::scale_size_manual(values = c(Car = 0.5, Lamp = 1.0)) + 
+#'   ggplot2::scale_linetype_manual(values = c(Car = "solid", Lamp = "dashed")) + 
 #'   ggplot2::labs(colour = ggplot2::element_blank(), 
 #'                 size = ggplot2::element_blank(),
 #'                 linetype = ggplot2::element_blank())
-# sensitivity_graphs <- function(.parametric_data = parametric_analysis(rebound_data, parameterization),
-#                                rebound_data, parameterization, 
-#                                x_var, 
-#                                y_vars = ReboundTools::rebound_terms,
-#                                linetype_var = ReboundTools::eeu_base_params$case, 
-#                                linecolour_var = ReboundTools::eeu_base_params$case, 
-#                                linesize_var = ReboundTools::eeu_base_params$case, 
-#                                graph_params = ReboundTools::sens_graph_params,
-#                                point_type_colname = ReboundTools::parametric_analysis_point_types$point_type_colname, 
-#                                sweep_points = ReboundTools::parametric_analysis_point_types$sweep, 
-#                                orig_points = ReboundTools::parametric_analysis_point_types$orig, 
-#                                Re_names = ReboundTools::graph_df_colnames$Re_names,
-#                                Re_values = ReboundTools::graph_df_colnames$Re_values) {
-#   
-#   y_var = match.arg(y_vars, several.ok = TRUE) %>% 
-#     unlist()
-#   
-#   p_data <- .parametric_data %>% 
-#     dplyr::select(tidyselect::all_of(x_var), tidyselect::all_of(y_vars), 
-#                   tidyselect::all_of(point_type_colname), tidyselect::all_of(linetype_var), 
-#                   tidyselect::all_of(linecolour_var), tidyselect::all_of(linesize_var)) %>% 
-#     tidyr::pivot_longer(cols = tidyselect::all_of(y_vars), names_to = Re_names, values_to = Re_values)
-#   
-#   line_data <- p_data %>% 
-#     dplyr::filter(.data[[ReboundTools::parametric_analysis_point_types$point_type_colname]] == sweep_points)
-#   point_data <- p_data %>% 
-#     dplyr::filter(.data[[ReboundTools::parametric_analysis_point_types$point_type_colname]] == orig_points)
-#   
-#   # Create the graph
-#   ggplot2::ggplot() + 
-#     ggplot2::geom_point(data = point_data, 
-#                         mapping = ggplot2::aes_string(x = x_var, y = Re_values, group = Re_names), 
-#                         colour = graph_params$orig_point_colour, 
-#                         size = graph_params$orig_point_size, 
-#                         shape = graph_params$orig_point_shape, 
-#                         stroke = graph_params$orig_point_stroke) +
-#     ggplot2::geom_line(data = line_data, 
-#                        mapping = ggplot2::aes_string(x = x_var, y = Re_values, group = Re_names, 
-#                                                      linetype = linetype_var, 
-#                                                      colour = linecolour_var, 
-#                                                      size = linesize_var))
-# }
+#'
+#' # Plot all rebound terms as a function of post-upgrade efficiency
+#' sens_params_3 <- list(Car = list(eta_engr_units_star = seq(35, 50, by = 0.5)), 
+#'                       Lamp = list(eta_engr_units_star = seq(70, 90, by = 5)))
+#' # Choose rebound terms to include in the graph and their order
+#' rebound_vars <- c("Re_dempl", "Re_emb", "Re_md", "Re_dsub", "Re_isub", 
+#'                   "Re_dinc", "Re_iinc", "Re_prod")
+#'                   
+#' sensitivity_graphs(rebound_data = df, parameterization = sens_params_3,
+#'                    x_var = "eta_engr_units_tilde", 
+#'                    y_var = rebound_vars) + 
+#'   ggplot2::facet_wrap(facets = "Case", scales = "free_x") +
+#'   ggplot2::scale_colour_manual(values = 
+#'         c(Re_dempl = ReboundTools::path_graph_params$dempl_colour,
+#'           Re_emb = ReboundTools::path_graph_params$emb_colour,
+#'           Re_md = ReboundTools::path_graph_params$md_colour, 
+#'           Re_dsub = ReboundTools::path_graph_params$dsub_colour,
+#'           Re_isub = ReboundTools::path_graph_params$isub_colour, 
+#'           Re_dinc = ReboundTools::path_graph_params$dinc_colour,
+#'           Re_iinc = ReboundTools::path_graph_params$iinc_colour,
+#'           Re_prod = ReboundTools::path_graph_params$prod_colour), 
+#'                                breaks = rebound_vars) +
+#'  ggplot2::scale_size_manual(values = 
+#'         c(Re_dempl = 0.2, 
+#'           Re_emb = ReboundTools::path_graph_params$emb_size,
+#'           Re_md = ReboundTools::path_graph_params$md_size, 
+#'           Re_dsub = ReboundTools::path_graph_params$dsub_size,
+#'           Re_isub = ReboundTools::path_graph_params$isub_size, 
+#'           Re_dinc = ReboundTools::path_graph_params$dinc_size,
+#'           Re_iinc = ReboundTools::path_graph_params$iinc_size,
+#'           Re_prod = ReboundTools::path_graph_params$prod_size), 
+#'                             breaks = rebound_vars) +
+#' ggplot2::scale_linetype_manual(values = 
+#'         c(Re_dempl = ReboundTools::path_graph_params$dempl_linetype, 
+#'           Re_emb = ReboundTools::path_graph_params$emb_linetype,
+#'           Re_md = ReboundTools::path_graph_params$md_linetype,
+#'           Re_dsub = ReboundTools::path_graph_params$dsub_linetype,
+#'           Re_isub = "11",
+#'           Re_dinc = ReboundTools::path_graph_params$dinc_linetype,
+#'           Re_iinc = "11",
+#'           Re_prod = ReboundTools::path_graph_params$prod_linetype), 
+#'                                breaks = rebound_vars) +
+#' ggplot2::labs(x = expression(tilde(eta)*" [mpg (Car) or lm/W (Lamp)]"), 
+#'               y = "Re terms [-]", 
+#'               colour = ggplot2::element_blank(),
+#'               size = ggplot2::element_blank(),
+#'               linetype = ggplot2::element_blank())
+sensitivity_graphs <- function(.parametric_data = parametric_analysis(rebound_data, parameterization),
+                               rebound_data = NULL, 
+                               parameterization = NULL,
+                               x_var,
+                               y_var,
+                               line_var = y_names_col,
+                               y_names_col = ReboundTools::graph_df_colnames$y_names_col,
+                               y_vals_col = ReboundTools::graph_df_colnames$y_vals_col,
+                               graph_params = ReboundTools::sens_graph_params,
+                               point_type_colname = ReboundTools::parametric_analysis_point_types$point_type_colname,
+                               sweep_points = ReboundTools::parametric_analysis_point_types$sweep,
+                               orig_points = ReboundTools::parametric_analysis_point_types$orig) {
+
+
+  p_data <- .parametric_data %>%
+    tidyr::pivot_longer(cols = tidyselect::all_of(y_var), names_to = y_names_col, values_to = y_vals_col) %>% 
+    # Arrange by the x variable so that all points (for geom_point) are in order.
+    dplyr::arrange(.data[[x_var]]) %>% 
+    dplyr::mutate(
+      # Set the factor levels so that the layering order will be 
+      # in the same order as the variables coming in here.
+      "{y_names_col}" := factor(.data[[y_names_col]], levels = y_var)
+    )
+
+  orig_data <- p_data %>%
+    dplyr::filter(.data[[ReboundTools::parametric_analysis_point_types$point_type_colname]] == orig_points)
+  line_data <- p_data %>%
+    dplyr::filter(.data[[ReboundTools::parametric_analysis_point_types$point_type_colname]] == sweep_points)
+
+  # Create the graph and return it
+  g <- ggplot2::ggplot()
+  if (graph_params$include_x_axis) {
+    g <- g +
+      ggplot2::geom_hline(yintercept = 0, size = 0.2)
+  }
+  g + 
+    ggplot2::geom_point(data = orig_data,
+                        # Pick up the dot colour from the line colour.
+                        mapping = ggplot2::aes_string(x = x_var, y = y_vals_col, colour = line_var),
+                        size = graph_params$orig_point_size,
+                        shape = graph_params$orig_point_shape,
+                        stroke = graph_params$orig_point_stroke,
+                        fill = graph_params$orig_point_fill) +
+    # Use geom_path so that we get nice rounded segments when using dashes.
+    ggplot2::geom_path(data = line_data,
+                       mapping = ggplot2::aes_string(x = x_var,
+                                                     y = y_vals_col,
+                                                     linetype = line_var,
+                                                     colour = line_var,
+                                                     size = line_var), 
+                       lineend = graph_params$lineend, 
+                       linejoin = graph_params$linejoin)
+}
+
+
+#' A sensitivity graph containing all rebound terms
+#' 
+#' Create a sensitivity graph with lines for all rebound terms. 
+#' 
+#' This function has the same arguments as `sensitivity_graphs()`,
+#' except that `yvar` is missing.
+#' `y_var` is missing, because the ordinate is assumed to be rebound values, 
+#' and all rebound components are included.
+#'
+#' @param .parametric_data A data frame, likely the result of calling `parametric_analysis()`.
+#'                         Default is `parametric_analysis(rebound_data, parameterization)`.
+#' @param rebound_data Rebound data, likely read by `load_eeu_data()`.
+#'                     Default is `NULL`.
+#' @param parameterization A list of lists that gives parameter sweeps.
+#'                         At the top level, the list items must be named for cases in `rebound_data`.
+#'                         At the next level, the parameters to be swept should be given, 
+#'                         along with their sweep values. 
+#'                         See examples.
+#'                         Default is `NULL`.
+#' @param x_var Strings that identify the x-axis and y-axis variables for this sensitivity graph.
+#'                    These variables must appear in `.parametric_data`.
+#'                    `x_var` must be a single string.
+#'                    `y_var` can be a vector of strings. 
+#'                    See examples.
+#' @param Re_terms A string vector that tells which rebound terms to include in the graph. 
+#'                 Default is `unlist(ReboundTools::rebound_terms)`.
+#' @param graph_params A list of parameters to control graph appearance. 
+#'                     See `ReboundTools::sens_graph_params`.
+#' @param point_type_colname,sweep_points,orig_points See `ReboundTools::parametric_analysis_point_types`.
+#' @param line_var The name of variable to be used to discriminate lines on the graph. 
+#'                 Default is `y_names_col`.
+#' @param y_names_col,y_vals_col See `ReboundTools::graph_df_colnames`.
+#' @param graph_params A list of parameters to control graph appearance. 
+#'                     See `ReboundTools::sens_graph_params`.
+#' @param point_type_colname,sweep_points,orig_points See `ReboundTools::parametric_analysis_point_types`.
+#'
+#' @return A ggplot2 graph showing sensitivity of all rebound terms to `x_var`.
+#'  
+#' @export
+#'
+#' @examples
+#' df <- load_eeu_data()
+#' sens_params <- list(Car = list(eta_engr_units_star = seq(35, 50, by = 0.5)), 
+#'                     Lamp = list(eta_engr_units_star = seq(70, 90, by = 5)))
+#' rebound_terms_graph(rebound_data = df, parameterization = sens_params, 
+#'                     x_var = "eta_engr_units_tilde") +
+#'   ggplot2::facet_wrap(facets = "Case", scales = "free_x")
+rebound_terms_graph <- function(.parametric_data = parametric_analysis(rebound_data, parameterization),
+                                rebound_data, 
+                                parameterization,
+                                x_var,
+                                Re_terms = unlist(ReboundTools::rebound_terms),
+                                line_var = y_names_col,
+                                y_names_col = ReboundTools::graph_df_colnames$y_names_col,
+                                y_vals_col = ReboundTools::graph_df_colnames$y_vals_col,
+                                graph_params = ReboundTools::sens_graph_params,
+                                point_type_colname = ReboundTools::parametric_analysis_point_types$point_type_colname,
+                                sweep_points = ReboundTools::parametric_analysis_point_types$sweep,
+                                orig_points = ReboundTools::parametric_analysis_point_types$orig) {
+  Re_terms <- match.arg(Re_terms, several.ok = TRUE)
+  
+  if (graph_params$use_latex_legend) {
+    legend_labs <- ReboundTools::latex_rebound_terms[Re_terms] %>% 
+      latex2exp::TeX() %>% 
+      unlist()
+  } else {
+    legend_labs <- Re_terms %>% unlist()
+  }
+  
+
+  sensitivity_graphs(.parametric_data = .parametric_data, 
+                     x_var = x_var, 
+                     y_var = Re_terms, 
+                     line_var = line_var, 
+                     y_vals_col = y_vals_col,
+                     y_names_col = y_names_col,
+                     graph_params = graph_params, 
+                     point_type_colname = point_type_colname,
+                     sweep_points = sweep_points,
+                     orig_points = orig_points) +
+    ggplot2::scale_colour_manual(values = c(Re_dempl = graph_params$dempl_colour,
+                                            Re_emb = graph_params$emb_colour,
+                                            Re_cap = graph_params$cap_colour,
+                                            Re_md = graph_params$md_colour, 
+                                            Re_empl = graph_params$empl_colour,
+                                            Re_dsub = graph_params$dsub_colour,
+                                            Re_isub = graph_params$isub_colour, 
+                                            Re_sub = graph_params$sub_colour,
+                                            Re_dinc = graph_params$dinc_colour,
+                                            Re_iinc = graph_params$iinc_colour,
+                                            Re_inc = graph_params$inc_colour,
+                                            Re_prod = graph_params$prod_colour,
+                                            Re_dir = graph_params$dir_colour,
+                                            Re_indir = graph_params$indir_colour,
+                                            Re_tot = graph_params$tot_colour), 
+                                 labels = legend_labs,
+                                 breaks = Re_terms) +
+    ggplot2::scale_size_manual(values = c(Re_dempl = graph_params$dempl_size,
+                                          Re_emb = graph_params$emb_size,
+                                          Re_md = graph_params$md_size,
+                                          Re_cap = graph_params$cap_size,
+                                          Re_empl = graph_params$empl_size,
+                                          Re_dsub = graph_params$dsub_size,
+                                          Re_isub = graph_params$isub_size,
+                                          Re_sub = graph_params$sub_size,
+                                          Re_dinc = graph_params$dinc_size,
+                                          Re_iinc = graph_params$iinc_size,
+                                          Re_inc = graph_params$inc_size,
+                                          Re_prod = graph_params$prod_size,
+                                          Re_dir = graph_params$dir_size,
+                                          Re_indir = graph_params$indir_size,
+                                          Re_tot = graph_params$tot_size),
+                               labels = legend_labs,
+                               breaks = Re_terms) +
+    ggplot2::scale_linetype_manual(values = c(Re_dempl = graph_params$dempl_linetype, 
+                                              Re_emb = graph_params$emb_linetype,
+                                              Re_cap = graph_params$cap_linetype,
+                                              Re_md = graph_params$md_linetype,
+                                              Re_empl = graph_params$empl_linetype,
+                                              Re_dsub = graph_params$dsub_linetype,
+                                              Re_isub = graph_params$isub_linetype,
+                                              Re_sub = graph_params$sub_linetype,
+                                              Re_dinc = graph_params$dinc_linetype,
+                                              Re_iinc = graph_params$iinc_linetype,
+                                              Re_inc = graph_params$inc_linetype,
+                                              Re_prod = graph_params$prod_linetype, 
+                                              Re_dir = graph_params$dir_linetype,
+                                              Re_indir = graph_params$indir_linetype,
+                                              Re_tot = graph_params$tot_linetype), 
+                                   labels = legend_labs,
+                                   breaks = Re_terms)
+}